requirements.txt
requests
mutagen
textblob
pydub
speechrecognition
langchain_google_genai
setuptools
transformers
python-dotenv
torch